Fix cache synchronize not clearing full length
If the address is not cache-line aligned, then the length it is
not aligned will not be properly cleared, because each iteration
assumes an entire cache line's worth has been cleared.

Instead, we first align to the next smallest cache line and add the
difference so we are sure to clear at least the full length
requested.